# .well-known/openapi.yaml
openapi: 3.1.0
info:
  title: Bybit Crypto Assistant API
  version: "1.0.0"
  description: API для просмотра портфеля Bybit, покупки криптовалют и расчёта PnL.
servers:
  - url: https://bybit-gpt-1.onrender.com
paths:
  /portfolio:
    get:
      summary: Текущий портфель (ненулевой баланс)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список активов
          content:
            application/json:
              schema:
                type: object
                properties:
                  portfolio:
                    type: array
                    items:
                      $ref: '#/components/schemas/PortfolioItem'
  /buy:
    post:
      summary: Купить криптовалюту по рыночной цене
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyRequest'
      responses:
        '200':
          description: Результат покупки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyResponse'
  /pnl:
    get:
      summary: PnL с учётом начального портфеля
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Значения PnL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PnLResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    PortfolioItem:
      type: object
      properties:
        coin:
          type: string
        balance:
          type: number
        locked:
          type: number
        usdValue:
          type: number
      required: [coin, balance]
    BuyRequest:
      type: object
      properties:
        symbol:
          type: string
        amount:
          type: number
      required: [symbol, amount]
    BuyResponse:
      type: object
      properties:
        order:
          type: object
          description: Детали ордера (ответ Bybit)
    PnLResponse:
      type: object
      properties:
        initial_value_usd:
          type: number
        current_value_usd:
          type: number
        pnl_usd:
          type: number
        pnl_percent:
          type: number